chore: rename color interface file
