+++ source/pixel/color/color.ts
