+++ source/pixel/color/spaces/hex.ts
@@ -1,7 +1,7 @@
 import {
   ColorSpace,
   IColor,
-} from '~pixel/color/icolor';
+} from '~pixel/color/color';
 
 import { Cmyk } from '~pixel/color/spaces/cmyk';
 import { Hsv } from '~pixel/color/spaces/hsv';
